refactor[VerificadorValidade]
